poc-agents/hotel_flight/pyproject.toml
[project]
name = "hotel-flight"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=1.2.6",
    "langchain-core>=1.2.7",
    "langchain-ollama>=1.0.1",
    "langgraph-supervisor>=0.0.31",
    "python-dotenv>=1.2.1",
]
